# openapi/kotona-api.json
{"openapi":"3.1.0","info":{"title":"KOTONA API Documentation","description":"일본 비즈니스 메일 분석 및 숙어 사전 서비스 API","version":"v1.0.0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"tags":[{"name":"Business Phrase","description":"비즈니스 일본어 숙어 라이브러리 API"},{"name":"Nuance Analysis","description":"일본어 비즈니스 문장 뉘앙스 분석 API"}],"paths":{"/analyze":{"post":{"tags":["Nuance Analysis"],"summary":"문장 뉘앙스 분석","description":"Gemini 로 본음/건전, 정중도 점수, 리스크, 추천 답장을 분석한다.","operationId":"analyze","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyzeRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NuanceResponseDTO"}}}}}}},"/api/phrases":{"get":{"tags":["Business Phrase"],"summary":"모든 숙어 조회","description":"DB에 저장된 모든 비즈니스 숙어를 정중도 순으로 조회","operationId":"getAllPhrases","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BusinessPhrase"}}}}}}}},"/api/phrases/search":{"get":{"tags":["Business Phrase"],"summary":"상황별 숙어 검색","description":"EMAIL, MEETING 등 특정 상황에 맞는 숙어만 필터링","operationId":"getPhrasesBySituation","parameters":[{"name":"situation","in":"query","description":"검색할 상황 태그(예: EMAIL, MEETING)","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BusinessPhrase"}}}}}}}},"/api/history":{"get":{"tags":["analysis-history-controller"],"operationId":"getAllHistory","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AnalysisHistory"}}}}}}}},"/api/health":{"get":{"tags":["health-check-controller"],"operationId":"healthCheck","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/history/{id}":{"delete":{"tags":["analysis-history-controller"],"operationId":"deleteHistory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"AnalyzeRequestDTO":{"type":"object","properties":{"text":{"type":"string"},"relationshipType":{"type":"string"}},"required":["text"]},"EvaluationDTO":{"type":"object","properties":{"summary":{"type":"string"},"keigo_check":{"type":"boolean"},"cushion_phrase_check":{"type":"boolean"}}},"FeedbackDTO":{"type":"object","properties":{"issues":{"type":"array","items":{"type":"string"}},"cultural_nuance":{"type":"string"}}},"HonneDTO":{"type":"object","properties":{"tatemae":{"type":"string"},"trueIntent":{"type":"string"},"actionItem":{"type":"string"}}},"MetricsDTO":{"type":"object","properties":{"politeness":{"type":"integer","format":"int32"},"indirectness":{"type":"integer","format":"int32"},"etiquette":{"type":"integer","format":"int32"}}},"NuanceResponseDTO":{"type":"object","properties":{"totalScore":{"type":"integer","format":"int32"},"category":{"type":"string"},"metrics":{"$ref":"#/components/schemas/MetricsDTO"},"evaluation":{"$ref":"#/components/schemas/EvaluationDTO"},"feedback":{"$ref":"#/components/schemas/FeedbackDTO"},"suggestions":{"type":"array","items":{"$ref":"#/components/schemas/SuggestionDTO"}},"sentiment":{"$ref":"#/components/schemas/SentimentDTO"},"riskAnalysis":{"$ref":"#/components/schemas/RiskAnalysisDTO"},"smartReplies":{"type":"array","items":{"$ref":"#/components/schemas/SmartReplyDTO"}}}},"RiskAnalysisDTO":{"type":"object","properties":{"riskLevel":{"type":"string"},"redFlags":{"type":"array","items":{"type":"string"}},"copingStrategy":{"type":"string"}}},"SentimentDTO":{"type":"object","properties":{"polarity":{"type":"string"},"confidence":{"type":"number","format":"double"},"honne":{"$ref":"#/components/schemas/HonneDTO"}}},"SmartReplyDTO":{"type":"object","properties":{"scenario":{"type":"string"},"content":{"type":"string"},"description":{"type":"string"},"nuanceLevel":{"type":"string"}}},"SuggestionDTO":{"type":"object","properties":{"text":{"type":"string"},"level":{"type":"string"}}},"BusinessPhrase":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"phrase":{"type":"string"},"meaning":{"type":"string"},"situation":{"type":"string"},"politenessLevel":{"type":"integer","format":"int32"},"usageExample":{"type":"string"}}},"AnalysisHistory":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userInput":{"type":"string"},"totalScore":{"type":"integer","format":"int32"},"category":{"type":"string"},"riskLevel":{"type":"string"},"fullAnalysisJson":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}}}}}